+++ src/xrGame/GametaskManager.cpp
@@ -105,8 +105,8 @@ CGameTask* CGameTaskManager::GiveGameTaskToActor(CGameTask* t, u32 timeToComplet
     GetGameTasks().push_back(SGameTaskKey(t->m_ID));
     GetGameTasks().back().game_task = t;
     t->m_ReceiveTime = Level().GetGameTime();
-    t->m_TimeToComplete = t->m_ReceiveTime + timeToComplete * 1000; // ms
-    t->m_timer_finish = t->m_ReceiveTime + timer_ttl * 1000; // ms
+    t->m_TimeToComplete = t->m_ReceiveTime + static_cast<u64>(timeToComplete) * 1000; // ms
+    t->m_timer_finish = t->m_ReceiveTime + static_cast<u64>(timer_ttl) * 1000; // ms
 
     std::stable_sort(GetGameTasks().begin(), GetGameTasks().end(), task_prio_pred);
