+++ src/xrGame/agent_enemy_manager.cpp
@@ -245,7 +245,7 @@ void CAgentEnemyManager::assign_enemies()
         (*I).m_probability *= 1.f - best;
 
         // recovering sort order
-        for (u32 i = 0, n = m_enemies.size() - 1; i < n; ++i)
+        for (size_t i = 0, n = m_enemies.size() - 1; i < n; ++i)
             if (m_enemies[i + 1] < m_enemies[i])
                 std::swap(m_enemies[i], m_enemies[i + 1]);
             else
