+++ src/xrGame/detail_path_manager_inline.h
@@ -43,7 +43,7 @@ IC const DetailPathManager::STravelPathPoint& CDetailPathManager::curr_travel_po
     return (m_path[curr_travel_point_index()]);
 }
 
-IC u32 CDetailPathManager::curr_travel_point_index() const
+IC size_t CDetailPathManager::curr_travel_point_index() const
 {
     VERIFY2(!m_path.empty() && (m_current_travel_point < m_path.size()),
         make_string("path[%d], current[%d]", m_path.size(), m_current_travel_point));
